pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "mlops_group35"
version = "0.0.1"
description = "DTU MLOps course project"
authors = [
  { name = "Group 35", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "dvc>=3.66.1",
    "fastapi==0.115.6",
    "numpy<2",
    "pyyaml>=6.0.3",
    "torch==2.0.1",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "hydra-core>=1.3.2",
    "wandb>=0.23.1",
    "pandas>=2.3.3",
    "scikit-learn>=1.8.0",
    "requests>=2.32.5",
    "streamlit>=1.53.0",
    "httpx>=0.28.1",
    "prometheus-client>=0.24.1",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "psutil>=7.2.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "kaggle>=1.8.3",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]

[project.scripts]
train = "src.mlops_group35.train:main"
